# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Praca inżynierska",
        "version": "1.0.0"
    },
    "paths": {
...
        "/channel/{id}": {
            "get": {
                "summary": "Konkretny kanał",
                "description": "Widok konkretnego kanału",
                "operationId": "b1a144d62d3b22db5a4f4a4fbb1b618c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id kanału",
                        "required": true,
                        "example": "7"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Otworzono widok kanału"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}